v2.0.111 TB Usenet Test.
TB Usenet Provider Tests

Technically works, Torrents can be added to queue and downloaded from TB.
- Min. required timeout in tests is 30s -> Probably safest sitting around 60.
- Files are initially added with name `NZB Upload`.
- No availability endpoint.
- No regex.

IMO, this should not be added as a seperate provider. TorBoxTorrentClient.cs should be adapted instead to work with both torrents and nzbs. Will probably require modifying client a little to allow for this... Also adding of NZB URLs similar to `magnet:xyz`